+++ 2 Business layer/CandidateEvaluator.Core/CandidateEvaluator.Core.csproj
