+++ 2 Business layer/CandidateEvaluator.Core/Dispatchers/Dispatcher.cs
@@ -2,12 +2,9 @@
 using CandidateEvaluator.Contract.Dispatchers;
 using CandidateEvaluator.Contract.Queries;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Services.Dispatchers
+namespace CandidateEvaluator.Core.Dispatchers
 {
     public class Dispatcher : IDispatcher
     {
