+++ 2 Business layer/CandidateEvaluator.Core/Dispatchers/QueryDispatcher.cs
@@ -1,14 +1,10 @@
 ﻿using CandidateEvaluator.Contract.Dispatchers;
 using CandidateEvaluator.Contract.Handlers;
-using CandidateEvaluator.Contract.Models;
 using CandidateEvaluator.Contract.Queries;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Services.Dispatchers
+namespace CandidateEvaluator.Core.Dispatchers
 {
     public class QueryDispatcher : IQueryDispatcher
     {
