+++ 2 Business layer/CandidateEvaluator.Core/Handlers/Commands/DeleteCategoryHandler.cs
@@ -1,14 +1,10 @@
 ﻿using CandidateEvaluator.Contract.Commands.Category;
 using CandidateEvaluator.Contract.Handlers;
-using CandidateEvaluator.Contract.Models;
 using CandidateEvaluator.Contract.Services;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Services.Handlers.Commands
+namespace CandidateEvaluator.Core.Handlers.Commands
 {
     public class DeleteCategoryHandler : ICommandHandler<DeleteCategory>
     {
