+++ 2 Business layer/CandidateEvaluator.Core/Handlers/Commands/UpdateCategoryHandler.cs
@@ -3,30 +3,28 @@ using CandidateEvaluator.Contract.Handlers;
 using CandidateEvaluator.Contract.Models;
 using CandidateEvaluator.Contract.Services;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace CandidateEvaluator.Services.Handlers.Commands
+namespace CandidateEvaluator.Core.Handlers.Commands
 {
-    public class CreateCategoryHandler : ICommandHandler<CreateCategory>
+    public class UpdateCategoryHandler : ICommandHandler<UpdateCategory>
     {
         private readonly ICategoryService _categoryService;
 
-        public CreateCategoryHandler(ICategoryService categoryService)
+        public UpdateCategoryHandler(ICategoryService categoryService)
         {
             _categoryService = categoryService;
         }
 
-        public async Task<Guid> HandleAsync(CreateCategory command)
+        public async Task<Guid> HandleAsync(UpdateCategory command)
         {
             var category = new Category
             {
+                Id = command.Id,
                 OwnerId = command.OwnerId,
                 Name = command.Name
             };
-            return await _categoryService.AddAsync(category);
+            return await _categoryService.Update(category);
         }
     }
 }
