+++ 3 Domain layer/CandidatesEvaluator.Contract/Queries/UserActivity/GetAllUserActivities.cs
@@ -1,0 +1,13 @@
+﻿using CandidateEvaluator.Contract.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CandidateEvaluator.Contract.Queries.UserActivity
+{
+    public class GetAllUserActivities : QueryBase, IQuery<List<RecentActivity>>
+    {
+    }
+}
